testing write on beta
