+++ public/db/counts.json
@@ -1,1 +1,1 @@
-{"count":0}
+{"count":3}
