+++ public/db/data.json
@@ -1,1 +1,1 @@
-{}
+{"2":[{"ownerID":"2","name":"percy jackson","page_num":"18","time_per_page":"6","rest_time":"1","reminder":"boop","book_count":1,"active":0},{"ownerID":"2","name":"nate","page_num":"1","time_per_page":"1","rest_time":"1","reminder":"ding","book_count":2,"active":1}],"3":[{"ownerID":"3","name":"the lightning thief","page_num":"1","time_per_page":"1","rest_time":"1","reminder":"ding","book_count":3,"active":1}]}
