+++ routes/home.js
@@ -2,7 +2,7 @@
 /*
  * GET home page.
  */
-var data = require('/app/public/db/data.json');
+var data = require('../public/db/data.json');
 
 exports.view = function(req, res){
 	var bookdat;
